--- a/tinygrad/tensor.py
+++ b/tinygrad/tensor.py
@@ -2266,7 +2266,7 @@ class Tensor:
     return self / (1 + self.abs())
 
   # ***** broadcasted elementwise ops *****
-  def _broadcast_to(self, shape:Tuple[sint, ...]) -> Tensor:
+  def _broadcast_to(self, shape:Tuple[sint, ...]):
     if self.shape == shape: return self
     reshape_arg, _ = _pad_left(self.shape, shape)
     if self.ndim > len(shape) or not all(sh in {s,1} or (s==0 and sh==1) for sh,s in zip(reshape_arg, shape)):
@@ -2282,13 +2282,18 @@ class Tensor:
       else: y_dtype = dtypes.from_py(y)
       if isinstance(y, Node): y = Tensor.from_node(y, device=self.device)
       else: y = Tensor(dtypes.as_const(y, y_dtype), self.device, y_dtype, requires_grad=False)
+
     if match_dtype and x.dtype != y.dtype:
       output_dtype = least_upper_dtype(x.dtype, y.dtype)
       x,y = x.cast(output_dtype), y.cast(output_dtype)
+
+    if reverse: x, y = y, x
+
+    # broadcast
     if x.shape != y.shape:
       out_shape = _broadcast_shape(x.shape, y.shape)
       x,y = x._broadcast_to(out_shape), y._broadcast_to(out_shape)
-    return (y,x) if reverse else (x,y)
+    return x,y
 
   def _to_const_val(self, x:Union[Tensor, ConstType]) -> Union[Tensor, ConstType]:
     # TODO: update with multi