`python -O test/external/external_test_speed_llama.py`

```
codegen(0)      mean runtime:  156.81ms, runs:   211.31,  127.76,  143.96,  137.93,  163.09
codegen(1)      mean runtime:  146.98ms, runs:   180.12,  137.88,  131.49,  145.81,  139.59
methodcache     mean runtime:   97.35ms, runs:    98.19,   92.33,  106.65,   95.18,   94.42
profile         mean runtime:  292.48ms, runs:   295.51,  290.46,  290.33,  291.24,  294.87
```


```
codegen(0)      mean runtime:  143.93ms, runs:   201.36,  118.72,  138.20,  128.39,  132.97
codegen(1)      mean runtime:  135.96ms, runs:   178.92,  114.88,  131.54,  123.23,  131.22
methodcache     mean runtime:   87.19ms, runs:    87.80,   97.08,   83.61,   83.77,   83.72
profile         mean runtime:  255.64ms, runs:   260.40,  254.79,  256.24,  253.32,  253.47
```