@roelofvandijk 2024-06-08
Cleaned up the diff. I needed the extra type annotation on `_broadcast_to` for mypy. 
I can remove the caches, however, they are responsible for a large part of the speedup. Should I?