@geohot 2024-06-08
It if trades complexity for speed, it's not worth it. As I have said many times, if it's not conceptually simpler (or at least the same), we don't want it.